fix: split out applyViteHtmlTransform and check extension
